Moar works

+++ user_data_service/database/password_protected.py
@@ -1,12 +1,26 @@
 from __future__ import print_function
 
+from functools import wraps
+
 import hashlib
 import uuid
 
 from database import Database
 
 
-class PasswordProtectedDatabase(Database):
+def password_protected(f):
+    @wraps(f)
+    def _password_protected(self, name, data):
+        password_information = self.database._get_password_information(name)
+        if not self.valid(data["password"], *password_information):
+            raise Exception("Permission Denied")
+
+        return f(self, name, data)
+
+    return _password_protected
+
+
+class PasswordProtected(Database):
     """Higher order database, takes a database and returns a password
     protected one.
 
@@ -18,10 +32,13 @@ class PasswordProtectedDatabase(Database):
     hash_ :: password, salt_value -> hash_value
         Generates the salted hash.
     """
-    def __init__(self,
+    def __init__(
+        self,
         database,
         salt=lambda: uuid.uuid4().hex,
-        hash_=lambda w, salt: hashlib.sha512(w + salt).hexdigest(),  # TODO use Passlib which uses PBKDF2 (rule #1 in security: never make your own)
+        hash_=lambda w, salt: hashlib.sha512(w + salt).hexdigest(),
+        # TODO use Passlib which uses PBKDF2
+        # (rule #1 in security: never make your own)
     ):
         self.database = database
         self.salt = salt
@@ -42,20 +59,25 @@ class PasswordProtectedDatabase(Database):
     def get(self, name):
         return self.database.get(name)
 
-    def delete(self, name):
-        # FIXME no password check..
+    @password_protected
+    def delete(self, name, data):
+        hash_value, salt_value = self.database._get_password_information(name)
+        if not self.valid(data["password"], hash_value, salt_value):
+            raise Exception("Permission Denied")
         return self.database.delete(name)
 
-    def update(self, data):
-        # FIXME no password check..
-        password_information(password, salt=lambda: salt_value) == self.database._get_password_informamtion(name)
-        return self.update(data)
+    @password_protected
+    def update(self, name, data):
+        return self.database.update(name, data)
 
-    def _get_password_informamtion(self, name):
-        return self.database._get_password_informamtion(name)
+    def connect(self):
+        return self.database.connect()
 
+    def close(self):
+        return self.database.close()
 
-    def password_information(self,
+    def password_information(
+        self,
         password,
         salt=None,
         hash_=None,
@@ -81,10 +103,12 @@ class PasswordProtectedDatabase(Database):
 
         return hash_value, salt_value
 
-
-    def validate(self,
+    def valid(
+        self,
         password,
         hash_value,
         salt_value
     ):
-        pass
+        print(self.hash_(password, salt_value))
+        print(hash_value)
+        return self.hash_(password, salt_value) == hash_value
